Extract documented classes to get rid of static
