+++ test-suite-groovy/src/test/groovy/io/micronaut/crac/OrderedResourceCheckpointSimulatorTest.groovy
@@ -2,13 +2,9 @@ package io.micronaut.crac
 
 import io.micronaut.context.BeanContext
 import io.micronaut.context.annotation.Property
-import io.micronaut.context.annotation.Requires
 import io.micronaut.crac.test.CheckpointSimulator
 import io.micronaut.test.extensions.spock.annotation.MicronautTest
 import jakarta.inject.Inject
-import jakarta.inject.Singleton
-import org.crac.Context
-import org.crac.Resource
 import spock.lang.Specification
 
 @MicronautTest
@@ -40,48 +36,4 @@ class OrderedResourceCheckpointSimulatorTest extends Specification {
         myBean.running
     }
     //end::test[]
-
-    @Requires(property = "spec.name", value = SPEC_NAME)
-    //tag::bean[]
-    @Singleton
-    static class ResourceBean {
-
-        private boolean running = true // <1>
-
-        boolean isRunning() {
-            return running
-        }
-
-        void stop() {
-            this.running = false
-        }
-
-        void start() {
-            this.running = true
-        }
-    }
-    //end::bean[]
-
-    @Requires(property = "spec.name", value = SPEC_NAME)
-    //tag::resource[]
-    @Singleton
-    static class ResourceBeanResource implements OrderedResource { // <1>
-
-        private final ResourceBean resourceBean
-
-        ResourceBeanResource(ResourceBean resourceBean) {
-            this.resourceBean = resourceBean
-        }
-
-        @Override
-        void beforeCheckpoint(Context<? extends Resource> context) throws Exception { // <2>
-            resourceBean.stop()
-        }
-
-        @Override
-        void afterRestore(Context<? extends Resource> context) throws Exception { // <3>
-            resourceBean.start()
-        }
-    }
-    //end::resource[]
 }
