+++ test-suite/src/test/java/io/micronaut/crac/ResourceBeanResource.java
@@ -1,0 +1,29 @@
+package io.micronaut.crac;
+
+import io.micronaut.context.annotation.Requires;
+import jakarta.inject.Singleton;
+import org.crac.Context;
+import org.crac.Resource;
+
+@Requires(property = "spec.name", value = OrderedResourceCheckpointSimulatorTest.SPEC_NAME)
+//tag::resource[]
+@Singleton
+class ResourceBeanResource implements OrderedResource { // <1>
+
+    private final ResourceBean resourceBean;
+
+    ResourceBeanResource(ResourceBean resourceBean) {
+        this.resourceBean = resourceBean;
+    }
+
+    @Override
+    public void beforeCheckpoint(Context<? extends Resource> context) throws Exception { // <2>
+        resourceBean.stop();
+    }
+
+    @Override
+    public void afterRestore(Context<? extends Resource> context) throws Exception { // <3>
+        resourceBean.start();
+    }
+}
+//end::resource[]
